# api/openapi/portal.yaml
openapi: "3.0.3"
info:
  title: Participant Portal API
  description: |
    Backend-for-frontend (BFF) API consumed exclusively by the Participant
    Portal React SPA. It is **not** the bank-to-switch payment API — that
    is `gateway.yaml`, uses mTLS, and speaks ISO 20022. This API serves
    *people* (browser sessions), not payment systems.

    ## Authentication — Authentik Proxy Provider

    All requests reach Portal API via the **Authentik Outpost** (reverse
    proxy). Authentik validates the session cookie, runs the authentication
    flow (username + password + TOTP) if there is no valid session, and
    injects identity headers before forwarding to Portal API.

    Portal API never validates sessions itself — it reads identity from
    headers the Authentik outpost injects:

    | Header                  | Source                        | Example                   |
    |-------------------------|-------------------------------|---------------------------|
    | `X-authentik-uid`       | Authentik user UUID           | `7f3a8c2d-...`            |
    | `X-authentik-username`  | Username (email)              | `alice@equity.ke`         |
    | `X-authentik-groups`    | Comma-separated group names   | `bank-admins,bank-viewers`|
    | `X-User-Role`           | Property mapping expression   | `BANK_ADMIN`              |
    | `X-Participant-Id`      | User attribute `participant_id`| `eqbl-ke-uuid` or `""`   |

...

    ## SPA auth lifecycle

...

...
    ## CSRF

...

    ## RBAC

...

...
servers:
...
security:
  - authentikSession: []
...
paths:
...
  /api/v1/csrf-token:
    get:
      summary: Issue a CSRF token
      description: |
        Returns a signed CSRF token bound to the current Authentik session.
        The SPA calls this once on startup and stores the token in memory.
        Include as `X-CSRF-Token` on all state-changing requests (POST,
        PUT, PATCH, DELETE). Tokens expire with the session.
      operationId: getCSRFToken
      tags: [Identity]
      responses:
        "200":
          description: CSRF token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Opaque signed token; treat as an opaque string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me:
    get:
      summary: Current user profile
      description: |
        Returns the caller's identity derived from Authentik-injected
        headers. The SPA calls this once on startup to populate its RBAC
        context (role, participantId, bankName).

        `bankName` is resolved from `X-Participant-Id` via the participant
        service; it is null for switch staff (`X-Participant-Id` is empty).
      operationId: getMe
      tags: [Identity]
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Banks
  # ---------------------------------------------------------------------------
  /api/v1/banks:
    get:
      summary: List participant banks
      description: |
        Returns all registered participant banks.
        **Roles**: `SWITCH_ADMIN`, `SWITCH_OPS`
      operationId: listBanks
      tags: [Banks]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/BankStatus"
          description: Filter by onboarding/operational status
      responses:
        "200":
          description: Paginated list of banks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      summary: Onboard a new participant bank
      description: |
        Creates a bank record and starts the onboarding state machine
        (APPLICATION → SANDBOX → CERTIFICATION → PRODUCTION_ACTIVE).
        **Roles**: `SWITCH_ADMIN`
      operationId: createBank
      tags: [Banks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankRequest"
      responses:
        "201":
          description: Bank created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: BIC already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/banks/{id}:
    get:
      summary: Get bank detail
      description: |
        **Roles**: `SWITCH_ADMIN`, `SWITCH_OPS` — any bank.
        `BANK_*` — own bank only (enforced via `X-Participant-Id`).
      operationId: getBank
      tags: [Banks]
      parameters:
        - $ref: "#/components/parameters/BankIdParam"
      responses:
        "200":
          description: Bank detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/banks/{id}/status:
    patch:
      summary: Update bank operational status
      description: |
        Transitions a bank to a new status. Invalid state-machine transitions
        return 422.
        **Roles**: `SWITCH_ADMIN`
      operationId: updateBankStatus
      tags: [Banks]
      parameters:
        - $ref: "#/components/parameters/BankIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBankStatusRequest"
      responses:
        "200":
          description: Bank after status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/banks/{id}/certificates:
    get:
      summary: List mTLS certificates for a bank
      description: |
        **Roles**: `SWITCH_ADMIN` — any bank. `BANK_ADMIN` — own bank only.
      operationId: listCertificates
      tags: [Banks]
      parameters:
        - $ref: "#/components/parameters/BankIdParam"
      responses:
        "200":
          description: Certificates registered for this bank
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      summary: Register a new certificate
      description: |
        Uploads a PEM-encoded X.509 certificate. The certificate's CN is
        used as the participant ID in the Kong mTLS header on the payment
        gateway service.
        **Roles**: `SWITCH_ADMIN`, `BANK_ADMIN` (own bank only)
      operationId: createCertificate
      tags: [Banks]
      parameters:
        - $ref: "#/components/parameters/BankIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCertificateRequest"
      responses:
        "201":
          description: Certificate registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/banks/{id}/certificates/{certId}:
    delete:
      summary: Revoke a certificate
      description: |
        Marks the certificate REVOKED. The payment gateway's Kong instance
        will reject future mTLS connections presenting this certificate.
        Action is written to the audit log.
        **Roles**: `SWITCH_ADMIN`, `BANK_ADMIN` (own bank only)
      operationId: revokeCertificate
      tags: [Banks]
      parameters:
        - $ref: "#/components/parameters/BankIdParam"
        - name: certId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Certificate revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/transactions:
    get:
      summary: List transactions
      description: |
        Paginated, filterable list sourced from the portal read projection
        (`transaction_view`).

        `BANK_*` roles: the `bic` parameter is **ignored** — the server
        always scopes to the BIC from `X-Participant-Id`. Providing a
        different value has no effect and is not an error.

        **Roles**: all authenticated roles.
      operationId: listTransactions
      tags: [Transactions]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/PaymentStatus"
        - name: bic
          in: query
          schema:
            type: string
          description: |
            Filter by source or destination BIC. Silently overridden to the
            caller's own BIC for `BANK_*` roles.
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Lower bound on `created_at` (RFC 3339, inclusive)
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: Upper bound on `created_at` (RFC 3339, inclusive)
        - name: minAmount
          in: query
          schema:
            type: integer
            minimum: 0
          description: Minimum amount in minor units (inclusive)
        - name: maxAmount
          in: query
          schema:
            type: integer
            minimum: 0
          description: Maximum amount in minor units (inclusive)
      responses:
        "200":
          description: Paginated transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/transactions/export:
    get:
      summary: Request a transaction export
      description: |
        Starts an async export job. Returns immediately with a `jobId`.
        Poll `GET /export-jobs/{jobId}` to check status and retrieve the
        download URL when the job completes. Accepts the same filters as
        `GET /transactions`.
        **Roles**: `SWITCH_ADMIN`, `SWITCH_OPS`, `BANK_ADMIN`, `BANK_OPERATOR`
      operationId: exportTransactions
      tags: [Transactions]
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, xlsx]
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/PaymentStatus"
        - name: bic
          in: query
          schema:
            type: string
          description: Silently overridden for BANK_* roles
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "202":
          description: Export job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJobAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/transactions/{paymentId}:
    get:
      summary: Get transaction detail
      description: |
        Returns full transaction detail including the embedded timeline.
        `BANK_*` roles are restricted to transactions where source or
        destination BIC matches their `X-Participant-Id`.
        **Roles**: all authenticated roles.
      operationId: getTransaction
      tags: [Transactions]
      parameters:
        - $ref: "#/components/parameters/PaymentIdParam"
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/transactions/{paymentId}/timeline:
    get:
      summary: Get transaction saga timeline
      description: |
        Ordered saga state transitions from `transaction_event_view`.
        Powers the `TransactionTimeline` stepper component in the portal.
        **Roles**: all authenticated roles (same scope as detail).
      operationId: getTransactionTimeline
      tags: [Transactions]
      parameters:
        - $ref: "#/components/parameters/PaymentIdParam"
      responses:
        "200":
          description: Ordered list of saga events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Export jobs
  # ---------------------------------------------------------------------------
  /api/v1/export-jobs/{jobId}:
    get:
      summary: Poll an export job
      description: |
        Returns the current status of an async export job. When `status` is
        `DONE`, `downloadUrl` contains a pre-signed URL valid for 15 minutes.
      operationId: getExportJob
      tags: [Transactions]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Export job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Dashboard
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    authentikSession:
      type: apiKey
      in: cookie
      name: authentik_session
      description: |
        httpOnly session cookie set by Authentik after successful
        authentication (username + password + TOTP). Managed entirely by
        the browser — the SPA does not read, store, or attach this cookie
        manually. Authentik's outpost validates it on every request and
        injects identity headers before forwarding to Portal API.

  # ---------------------------------------------------------------------------
  # Reusable parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page number
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...